requirements.txt
Django==1.5.4
Fabric==1.7.0
MySQL-python==1.2.4
South==0.8.2
amqp==1.0.13
anyjson==0.3.3
argparse==1.2.1
billiard==2.7.3.32
celery==3.0.23
django-braces==1.2.2
django-celery==3.0.23
django-extensions==1.2.1
django-frontend==0.3.1
django-frontend-skeleton==2.0.0
django-suit==0.2.4
gdata==2.0.18
gunicorn==18.0
kombu==2.5.14
paramiko==1.11.0
pycrypto==2.6
python-dateutil==2.1
pytz==2013d
six==1.4.1
wsgiref==0.1.2
